keyring_test.go: various test clean-ups
Instead of repeating function calls, it's better to declare the test cases as a
data input.  Iterate over them, and while we're there, un-export the test
helper function and make the failure functions more consistent.  This all reveals a
bug;

	$ go test -v . -run TestBasic
	=== RUN   TestBasicSetGet
	--- FAIL: TestBasicSetGet (0.32s)
		keyring_test.go:29: expected 'cr4zyp!s\%' for alice, got 'cr4zyp!s\134%'
	FAIL
	exit status 1

Add a test case for every ASCII to show that this only affects backslash.