keyring_darwin.go: handle funky passwords (#14)
* keyring_darwin.go: handle funky passwords

The 'security' command adds a hex representation of the password if it contains
certain characters in it.  Without the updated Regex, it fails to retrieve the
password, and the test fails with:

	$ go test .
	Get() error: keyring: Password not found
	--- FAIL: TestBasicSetGet (0.21s)
		keyring_test.go:24: Get() error for alice: keyring: Password not found
	FAIL
	FAIL	src/github.com/tmc/keyring	0.217s

* keyring_test.go: various test clean-ups

Instead of repeating function calls, it's better to declare the test cases as a
data input.  Iterate over them, and while we're there, un-export the test
helper function and make the failure functions more consistent.  This all reveals a
bug;

	$ go test -v . -run TestBasic
	=== RUN   TestBasicSetGet
	--- FAIL: TestBasicSetGet (0.32s)
		keyring_test.go:29: expected 'cr4zyp!s\%' for alice, got 'cr4zyp!s\134%'
	FAIL
	exit status 1

Add a test case for every ASCII to show that this only affects backslash.

* keyring_darwin: un-escape backslashes

It seems that the keyring escapes backslashes in its output; if that happens,
unescape them.  Hopefully there are no versions of the keyring that don't do
this, being used by people with '\134' in their password.

* keyring_test: un-break initial test run

The tests leave stale state in the keyring file between runs so I didn't notice
this until I added another test case and it failed.  Restore the first Get()
error branch to warn via testing.Logf()

* keyring_darwin.go: support unicode passwords

Probably supporting backslash alone is not sufficient for the variety
of passwords people might conceivably use.